PROGRESSA-1: Split models in individuals files
