+++ src/app/progressa/shared/models/exercise-set.model.ts
@@ -1,0 +1,7 @@
+export interface ExerciseSet {
+  id: number;
+  number: number;
+  weight: number;
+  repetitions: number;
+  rir: number;
+}
