+++ src/app/progressa/shared/models/exercise.model.ts
@@ -1,0 +1,7 @@
+import { ExerciseSet } from "./exercise-set.model";
+
+export interface Exercise {
+  id: number;
+  name: string;
+  sets: ExerciseSet[];
+}
